sdk-metrics: Apply attributes advice to the custom view
This change fixes an issue with unexpected metric attributes from users registeredView.

I use the following code to change aggregation for all histograms to base2ExponentialBucket:
```
        return provider.registerView(
                InstrumentSelector.builder().setType(InstrumentType.HISTOGRAM).build(),
                View.builder().setAggregation(Aggregation.base2ExponentialBucketHistogram()).build()
        );
```

But after that I get high cardinality tags like `url.full` produced for http metrics.
This happens because the view above uses `AttributesProcessor.noop()` and all span attributes added to the metric.
But it is not an expected behaviour. I still want the default metric advice attributes to be applied and `url.full` to be filtered out.